added error message
