+++ app/signup/page.tsx
@@ -2,7 +2,6 @@ import Link from 'next/link'
 import { headers, cookies } from 'next/headers'
 import { createClient } from '@/utils/supabase/server'
 import { redirect } from 'next/navigation'
-import AuthButton from '@/components/AuthButton'
 import UpdateForm from '@/components/form/UpdateProfile'
 
 export default async function Login({
